nodeice-matrix/requirements.txt
pillow>=9.0.0
pyyaml>=6.0
